Don't create intermediate maps when sending single form entry
Previously all form posting was done via the Map parameter
version which created an intermediate singleton map for
each other call to form taking either an entry or a name/value
combination.

This was unnecessary and now the main form sending method
is the name/value pair version which all overloaded versions
forward to.